pyproject.toml
[tool.poetry]
name = "bd_project"
version = "0.1.0"
description = ""
authors = ["Braindead3 <malosay21@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
WTForms = "^3.0.0"
Flask-Login = "^0.5.0"
Flask-WTF = "^1.0.0"
peewee = "^3.14.8"
Flask-Bcrypt = "^0.7.1"
email-validator = "^1.1.3"
phonenumbers = "^8.12.36"
gunicorn = "^20.1.0"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
